GravityCalculator
For your unit 1 assignment, you must first install the Netbeans development and the Java Software Development Kit and test the installation. Detailed instructions for doing this are provided in the unit overview. Students who are using the virtual computing lab will not be required to install the software as this has already been completed. Students who will be using mobile computing devices should consider creating a free account at www.programr.com and creating a Java project in which to complete this (and future) assignments.

In this assignment, you will create a program that computes the distance an object will fall in Earth's gravity.

1. Create a new class called GravityCalculator.
2. Copy and paste the following initial version:
3. Run it
4. Modify the example program to compute the position of an object after falling for 10 seconds, outputting the position in meters. The formula in Math notation is:
x(t) = 0.5at^2 + vit + xi

Variable	Meaning	Value
a	Acceleration (m/s2)	-9.81
t	Time (s)	10
vi	Initial velocity (m/s)	0
xi	Initial position	1000

Submit your GravityCalculator.java file along with the output you received when running it. The output should be captured as a screen capture or by cutting and pasting the output from the window that the output of your program is displayed in.